Move shadcn UI to packages/ui
