+++ apps/workbench/package.json
@@ -18,19 +18,16 @@
     "test:watch": "vitest"
   },
   "dependencies": {
-    "@radix-ui/react-slot": "^1.2.4",
+    "@acp/ui": "workspace:*",
     "@tanstack/react-query": "^5.90.0",
     "@tauri-apps/api": "^2.0.0",
     "@tauri-apps/plugin-dialog": "^2.4.0",
     "@vitejs/plugin-react": "^5.0.0",
-    "class-variance-authority": "^0.7.1",
-    "clsx": "^2.1.1",
     "lucide-react": "^0.468.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-markdown": "^10.1.0",
     "remark-gfm": "^4.0.1",
-    "tailwind-merge": "^3.5.0",
     "vite": "^7.0.0",
     "zustand": "^5.0.0"
   },
