+++ apps/workbench/src/shared/lib/index.ts
@@ -1,2 +1,1 @@
 export { stripAnsi } from "./ansi";
-export { cn } from "./utils";
