+++ apps/workbench/src/widgets/follow-up-composer/FollowUpComposer.tsx
@@ -11,7 +11,7 @@ import {
   PromptInputAction,
   PromptInputActions,
   PromptInputTextarea,
-} from "../../shared/ui";
+} from "@acp/ui";
 
 type FollowUpComposerProps = {
   value: string;
