--- packages/.gitkeep
@@ -1,1 +1,0 @@
-# Reserved for future shared workspace packages.
