+++ packages/ui/src/Card.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import { cn } from "../lib/utils";
+import { cn } from "./lib/utils";
 
 export const cardClassName =
   "rounded-lg border bg-card text-card-foreground shadow-card";
