+++ packages/ui/src/Message.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import { cn } from "../lib/utils";
+import { cn } from "./lib/utils";
 
 type MessageProps = React.HTMLAttributes<HTMLElement> & {
   as?: "article" | "div";
