+++ packages/ui/src/PromptInput.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import { cn } from "../lib/utils";
+import { cn } from "./lib/utils";
 import { Textarea } from "./Textarea";
 
 type PromptInputContextValue = {
